# vulkan: further optimize mul_mat_vec using larger loads