There are a few things in this PR:
- Use pipeline_robustness to disable bounds-checking for some pipelines (reduces integer instruction load).
- Apply the same kind of optimizations I did to Q4_K in https://github.com/ggerganov/llama.cpp/pull/10296 to Q5_K/Q6_K.
- Add vec4 dequant functions that use 16b loads and use them to do 8 K values per thread per iteration in mul_mat_vec, to help reduce load on the memory system.

Some performance results on RTX 4070. Note that the directed tests tend to fit in L2 on this system so they are much less memory-limited and don't reflect the real performance gain in networks (which tend to be framebuffer-limited, and benefit less from these optimizations). Also, this "Q4_K" network I tested uses all of Q4_K/Q5_K/Q6_K and is showing some benefit from the optimizations on those shaders (i.e. it's not all from the robustness change).

```
before:
  MUL_MAT(type_a=f32,type_b=f32,m=4096,n=1,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3]):                   11076 runs -   484.48 us/run - 117.44 MFLOP/run - 242.40 GFLOPS
  MUL_MAT(type_a=f16,type_b=f32,m=4096,n=1,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3]):                   20448 runs -   247.18 us/run - 117.44 MFLOP/run - 475.11 GFLOPS
  MUL_MAT(type_a=q4_0,type_b=f32,m=4096,n=1,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3]):                  65604 runs -    76.82 us/run - 117.44 MFLOP/run -   1.53 TFLOPS
  MUL_MAT(type_a=q4_1,type_b=f32,m=4096,n=1,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3]):                  58788 runs -    86.02 us/run - 117.44 MFLOP/run -   1.37 TFLOPS
  MUL_MAT(type_a=q5_0,type_b=f32,m=4096,n=1,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3]):                  36636 runs -   137.31 us/run - 117.44 MFLOP/run - 855.27 GFLOPS
  MUL_MAT(type_a=q5_1,type_b=f32,m=4096,n=1,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3]):                  34080 runs -   147.13 us/run - 117.44 MFLOP/run - 798.20 GFLOPS
  MUL_MAT(type_a=q8_0,type_b=f32,m=4096,n=1,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3]):                  36636 runs -   139.35 us/run - 117.44 MFLOP/run - 842.79 GFLOPS
  MUL_MAT(type_a=q2_K,type_b=f32,m=4096,n=1,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3]):                  51120 runs -    97.95 us/run - 117.44 MFLOP/run -   1.20 TFLOPS
  MUL_MAT(type_a=q3_K,type_b=f32,m=4096,n=1,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3]):                  40044 runs -   126.78 us/run - 117.44 MFLOP/run - 926.32 GFLOPS
  MUL_MAT(type_a=q4_K,type_b=f32,m=4096,n=1,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3]):                  88608 runs -    56.83 us/run - 117.44 MFLOP/run -   2.07 TFLOPS
  MUL_MAT(type_a=q5_K,type_b=f32,m=4096,n=1,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3]):                  40896 runs -   123.38 us/run - 117.44 MFLOP/run - 951.87 GFLOPS
  MUL_MAT(type_a=q6_K,type_b=f32,m=4096,n=1,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3]):                  40896 runs -   123.12 us/run - 117.44 MFLOP/run - 953.83 GFLOPS
  MUL_MAT(type_a=iq4_nl,type_b=f32,m=4096,n=1,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3]):                11076 runs -   466.42 us/run - 117.44 MFLOP/run - 251.79 GFLOPS
  
| phi3 3B Q4_K - Medium          |   2.23 GiB |     3.82 B | Vulkan     | 100 |         tg128 |        105.07  0.19 |
| baichuan 13B Q4_0              |   7.44 GiB |    13.90 B | Vulkan     | 100 |         tg128 |         39.87  0.41 |
| starcoder2 7B Q4_0             |   3.76 GiB |     7.17 B | Vulkan     | 100 |         tg128 |         65.04  0.23 |
| llama 3B Q8_0                  |   3.18 GiB |     3.21 B | Vulkan     | 100 |         tg128 |         90.59  0.55 |

after:
  MUL_MAT(type_a=f32,type_b=f32,m=4096,n=1,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3]):                   10224 runs -   492.94 us/run - 117.44 MFLOP/run - 238.24 GFLOPS
  MUL_MAT(type_a=f16,type_b=f32,m=4096,n=1,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3]):                   20448 runs -   250.89 us/run - 117.44 MFLOP/run - 468.10 GFLOPS
  MUL_MAT(type_a=q4_0,type_b=f32,m=4096,n=1,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3]):                  97980 runs -    51.10 us/run - 117.44 MFLOP/run -   2.30 TFLOPS
  MUL_MAT(type_a=q4_1,type_b=f32,m=4096,n=1,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3]):                  98832 runs -    50.76 us/run - 117.44 MFLOP/run -   2.31 TFLOPS
  MUL_MAT(type_a=q5_0,type_b=f32,m=4096,n=1,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3]):                  61344 runs -    81.83 us/run - 117.44 MFLOP/run -   1.44 TFLOPS
  MUL_MAT(type_a=q5_1,type_b=f32,m=4096,n=1,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3]):                  51120 runs -    99.18 us/run - 117.44 MFLOP/run -   1.18 TFLOPS
  MUL_MAT(type_a=q8_0,type_b=f32,m=4096,n=1,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3]):                  37488 runs -   135.20 us/run - 117.44 MFLOP/run - 868.67 GFLOPS
  MUL_MAT(type_a=q2_K,type_b=f32,m=4096,n=1,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3]):                  51972 runs -    96.68 us/run - 117.44 MFLOP/run -   1.21 TFLOPS
  MUL_MAT(type_a=q3_K,type_b=f32,m=4096,n=1,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3]):                  44304 runs -   113.30 us/run - 117.44 MFLOP/run -   1.04 TFLOPS
  MUL_MAT(type_a=q4_K,type_b=f32,m=4096,n=1,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3]):                  94572 runs -    52.91 us/run - 117.44 MFLOP/run -   2.22 TFLOPS
  MUL_MAT(type_a=q5_K,type_b=f32,m=4096,n=1,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3]):                  48564 runs -   104.13 us/run - 117.44 MFLOP/run -   1.13 TFLOPS
  MUL_MAT(type_a=q6_K,type_b=f32,m=4096,n=1,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3]):                  47712 runs -   106.42 us/run - 117.44 MFLOP/run -   1.10 TFLOPS
  MUL_MAT(type_a=iq4_nl,type_b=f32,m=4096,n=1,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3]):                11076 runs -   457.97 us/run - 117.44 MFLOP/run - 256.43 GFLOPS

| phi3 3B Q4_K - Medium          |   2.23 GiB |     3.82 B | Vulkan     | 100 |         tg128 |        109.65  0.33 |
| baichuan 13B Q4_0              |   7.44 GiB |    13.90 B | Vulkan     | 100 |         tg128 |         46.00  0.50 |
| starcoder2 7B Q4_0             |   3.76 GiB |     7.17 B | Vulkan     | 100 |         tg128 |         74.85  1.05 |
| llama 3B Q8_0                  |   3.18 GiB |     3.21 B | Vulkan     | 100 |         tg128 |         94.51  1.52 |
```

- [x] I have read the [contributing guidelines](https://github.com/ggerganov/llama.cpp/blob/master/CONTRIBUTING.md)
- Self-reported review complexity:
  - [ ] Low
  - [ ] Medium
  - [x] High