@netrunnereve 2024-11-19
Wow this is really nice. Here are my numbers on the RX 570 for this PR.

Before:
| model                          |       size |     params | backend    | ngl | threads |          test |                  t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ------: | ------------: | -------------------: |
| llama 8B Q4_0                  |   4.33 GiB |     8.03 B | Vulkan     | 100 |       8 |         tg128 |         11.61 ± 0.05 |
| llama 8B Q8_0                 |   7.95 GiB |     8.03 B | Vulkan     | 100 |       8 |         tg128 |          6.60 ± 0.02 |
| llama 8B Q4_K - Small          |   4.36 GiB |     8.03 B | Vulkan     | 100 |       8 |         tg128 |          9.18 ± 0.02 |

After:
| model                          |       size |     params | backend    | ngl | threads |          test |                  t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ------: | ------------: | -------------------: |
| llama 8B Q4_0                  |   4.33 GiB |     8.03 B | Vulkan     | 100 |       8 |         tg128 |         16.66 ± 0.05 |
| llama 8B Q8_0                  |   7.95 GiB |     8.03 B | Vulkan     | 100 |       8 |         tg128 |         11.67 ± 0.00 |
| llama 8B Q4_K - Small          |   4.36 GiB |     8.03 B | Vulkan     | 100 |       8 |         tg128 |          9.62 ± 0.01 |

There isn't that much change with the K-quants but Q8 now runs faster than the old Q4! :tada: 

Bonus results:
| model                          |       size |     params | backend    | ngl | threads |          test |                  t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ------: | ------------: | -------------------: |
| llama 8B Q4_0 (8 rows/workgroup)                 |   4.33 GiB |     8.03 B | Vulkan     | 100 |       8 |         tg128 |         18.23 ± 0.05 |
| llama 8B Q4_0 (8 rows + move out delta)                 |   4.33 GiB |     8.03 B | Vulkan     | 100 |       8 |         tg128 |         19.66 ± 0.02 |

As discussed in #10296 calculating 8 rows at a time helps with the 570. From what I've read online AMD GCN apparently likes longer running shaders. If I optimize the delta calculation using the code below I nearly get 20t/s - and that's all I can come up with for tonight.

``` diff
------------ ggml/src/ggml-vulkan/vulkan-shaders/dequant_funcs.comp ------------
index 5fc1ba4a..7f3cb33a 100644
@@ -30,9 +30,9 @@ vec2 dequantize(uint ib, uint iqs, uint a_offset) {
     return (vec2(vui & 0xF, vui >> 4) - 8.0f) * d;
 }
 vec4 dequantize4(uint ib, uint iqs, uint a_offset) {
-    const float d = float(data_a_packed16[a_offset + ib].d);
+    //const float d = float(data_a_packed16[a_offset + ib].d);
     const uint vui = uint(data_a_packed16[a_offset + ib].qs[iqs/2]);
-    return (vec4(vui & 0xF, (vui >> 4) & 0xF, (vui >> 8) & 0xF, (vui >> 12) & 0xF) - 8.0f) * d;
+    return (vec4(vui & 0xF, (vui >> 4) & 0xF, (vui >> 8) & 0xF, (vui >> 12) & 0xF) - 8.0f);
 }
 #endif
 
@@ -95,10 +95,10 @@ vec2 dequantize(uint ib, uint iqs, uint a_offset) {
     return vec2(int(data_a[a_offset + ib].qs[iqs]), int(data_a[a_offset + ib].qs[iqs + 1])) * d;
 }
 vec4 dequantize4(uint ib, uint iqs, uint a_offset) {
-    const float d = float(data_a_packed16[a_offset + ib].d);
+    //const float d = float(data_a_packed16[a_offset + ib].d);
     uint32_t v0 = data_a_packed16[a_offset + ib].qs[iqs/2];
     uint32_t v1 = data_a_packed16[a_offset + ib].qs[iqs/2 + 1];
-    return vec4(int8_t(v0 & 0xFF), int8_t((v0 >> 8) & 0xFF), int8_t(v1 & 0xFF), int8_t((v1 >> 8) & 0xFF)) * d;
+    return vec4(int8_t(v0 & 0xFF), int8_t((v0 >> 8) & 0xFF), int8_t(v1 & 0xFF), int8_t((v1 >> 8) & 0xFF));
 }
 #endif
 

------------- ggml/src/ggml-vulkan/vulkan-shaders/mul_mat_vec.comp -------------
index 00807a06..6a15bce5 100644
@@ -73,16 +73,22 @@ void iter(inout FLOAT_TYPE temp[NUM_ROWS], const uint first_row, const uint num_
 #if K_PER_ITER == 8
         const vec4 v = dequantize4(ib, iqs, a_offset);
         const vec4 v2 = dequantize4(ib, iqs+(4/QUANT_R), a_offset);
+        FLOAT_TYPE rowtmp = 0;
 
         // matrix multiplication
-        temp[n] = fma(FLOAT_TYPE(v.x), b0, temp[n]);
-        temp[n] = fma(FLOAT_TYPE(v.y), b1, temp[n]);
-        temp[n] = fma(FLOAT_TYPE(v.z), b2, temp[n]);
-        temp[n] = fma(FLOAT_TYPE(v.w), b3, temp[n]);
-        temp[n] = fma(FLOAT_TYPE(v2.x), b4, temp[n]);
-        temp[n] = fma(FLOAT_TYPE(v2.y), b5, temp[n]);
-        temp[n] = fma(FLOAT_TYPE(v2.z), b6, temp[n]);
-        temp[n] = fma(FLOAT_TYPE(v2.w), b7, temp[n]);
+        rowtmp = FLOAT_TYPE(v.x) * b0;
+        rowtmp = fma(FLOAT_TYPE(v.y), b1, rowtmp);
+        rowtmp = fma(FLOAT_TYPE(v.z), b2, rowtmp);
+        rowtmp = fma(FLOAT_TYPE(v.w), b3, rowtmp);
+        rowtmp = fma(FLOAT_TYPE(v2.x), b4, rowtmp);
+        rowtmp = fma(FLOAT_TYPE(v2.y), b5, rowtmp);
+        rowtmp = fma(FLOAT_TYPE(v2.z), b6, rowtmp);
+        rowtmp = fma(FLOAT_TYPE(v2.w), b7, rowtmp);
+#if defined(DATA_A_Q4_0) || defined(DATA_A_Q8_0)
+        const float d = float(data_a_packed16[a_offset + ib].d);
+        rowtmp *= d;
+#endif
+        temp[n] += rowtmp;
 #else
         const vec2 v = dequantize(ib, iqs, a_offset);
```