@0cc4m 2024-11-19
I'll add some model benchmarks later, but here's the test-backend-ops perf results.

Looks like a significant step forward, most effective on Nvidia (not surprising since that's what you're working on), but also significantly positive on AMD. I'm not sure why the difference between AMD Vulkan and ROCm is larger than Nvidia Vulkan and CUDA.

My Intel A770 also really liked the changes in q4_0, q4_1, q5_0 and q5_1, but not q8_0 and q6_k, not sure what's going on there.

![NVIDIA_GeForce_RTX_3090_performance_comparison](https://github.com/user-attachments/assets/18e38fa1-61c6-420e-86bc-1f569f1e7259)
![AMD_Radeon_RX_6800_XT_performance_comparison](https://github.com/user-attachments/assets/2808334a-b179-4b50-b9f4-d727bb56462f)
![AMD_Radeon_(TM)_Pro_VII_performance_comparison](https://github.com/user-attachments/assets/ba824a81-1b24-48d9-9b9b-57ec7c0178d8)
![Intel(R)_Arc(tm)_A770_Graphics_(DG2)_performance_comparison](https://github.com/user-attachments/assets/cd8bbb4c-6870-4307-b298-a6dcce659166)