@0cc4m 2024-11-20
Here's the results for Llama3 8B q4_0 and q4_k_s and Mistral Nemo q6_k. Looks good. The difference to CUDA/ROCm is larger than I remembered. Intel didn't like the change in some areas, so at a later point we might want to make the `K_PER_ITER` thing a specialization constant.

![mmv2_AMD_Radeon_RX_6800_XT_performance_comparison](https://github.com/user-attachments/assets/7fc60128-e3d9-4f7a-9ed5-df48f179242c)
![mmv2_AMD_Radeon_(TM)_Pro_VII_performance_comparison](https://github.com/user-attachments/assets/41101d5c-6e5c-4fc2-8af9-eddcd5e017f8)
![mmv2_Intel(R)_Arc(tm)_A770_Graphics_(DG2)_(Intel_open-source_Mesa_driver)_performance_comparison](https://github.com/user-attachments/assets/7c395453-9405-4ede-aaea-d83260733648)
![mmv2_NVIDIA_GeForce_RTX_3090_performance_comparison](https://github.com/user-attachments/assets/1faed353-b434-45b9-a903-7a44e7aaa604)